# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API for job posting website",
        "description": "API for website that provides job vacancies in various IT companies all over Belarus. This web-application is ment for students as well as for people with no (or little) working experience in IT sphere, as it offers exactly those positions, where no experience is required.",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/authenticate": {
            "get": {
                "tags": [
                    "authenticate"
                ],
                "summary": "Get info about user",
                "description": "Getting information about the user",
                "operationId": "getUserInfo",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "name": "nikasd",
                                    "email": "asdqwee@mail.ru",
                                    "registered_at": "2017-07-31 11:33:07"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": "400",
                                    "title": "Invalid token",
                                    "detail": "Could not decode token: The token \"eyJ0eXAiOiJKV1QiLCJhsbGciOiJIUzI1NiJ9.eyJleHAiOjE1MTM2NzIwNDgsInN1YiI6NiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdC9hcGkvYXV0aGVudGljYXRlIiwiaWF0IjoxNTAxNTc2MDQ5LCJuYmYiOjE1MDE1NzYwNDksImp0aSI6IkZVcW42NXRlY0xDSm5kN1AifQ.K2wOp6Qk0URWaYV_FXxTqWQPx-zyyzglVryMpvLYZgU\" is an invalid JWS"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": "400",
                                    "title": "Invalid token",
                                    "detail": "The token has been blacklisted"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "authenticate"
                ],
                "summary": "Login",
                "description": "Authenticate user with email and password",
                "operationId": "getToken",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "User email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "User password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "remember",
                        "in": "formData",
                        "description": "Remember user",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MTM2ODI1MDQsInN1YiI6NiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdC9hcGkvYXV0aGVudGljYXRlIiwiaWF0IjoxNTAxNTg2NTA0LCJuYmYiOjE1MDE1ODY1MDQsImp0aSI6ImJvM3E3OG9uQjEySE1ZU0oifQ.hQNn9CYY0GHvDZ2xCC3KD6kOM6tQllQpieyx4bSOauk"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Access is denied",
                        "examples": [
                            {
                                "application/json": {
                                    "errors": {
                                        "status": "401",
                                        "title": "Access is denied",
                                        "detail": "Wrong password",
                                        "source": {
                                            "parameter": "http:\\/\\/localhost\\/api\\/authenticate"
                                        }
                                    }
                                }
                            },
                            {
                                "application/json": {
                                    "errors": {
                                        "status": "401",
                                        "title": "Access is denied",
                                        "detail": "Wrong email",
                                        "source": {
                                            "parameter": "http:\\/\\/localhost\\/api\\/authenticate"
                                        }
                                    }
                                }
                            }
                        ]
                    },
                    "500": {
                        "description": "Undeclared error",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": "500",
                                    "title": "Undeclared error",
                                    "detail": "Something went wrong!",
                                    "source": {
                                        "parameter": "http:\\/\\/localhost\\/api\\/authenticate"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/authenticate/{token}": {
            "delete": {
                "tags": [
                    "authenticate"
                ],
                "summary": "Logout",
                "description": "Brings a token to the black list",
                "operationId": "logoutUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "examples": {
                            "application/json": {
                                "name": "nikasd",
                                "email": "asdqwee@mail.ru"
                            }
                        }
                    },
                    "401": {
                        "description": "Token in the blacklist",
                        "examples": {
                            "application/json": {
                                "error": {
                                    "status": 401,
                                    "title": "Bad Request",
                                    "detail": "The token has been blacklisted",
                                    "source": {
                                        "parameter": "http://localhost/api/authenticate/eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjYsImlzcyI6Imh0dHA6Ly9sb2NhbGhvc3QvYXBpL2F1dGhlbnRpY2F0ZSIsImlhdCI6MTUwMTU5MDEzOCwiZXhwIjoxNTAxNTkzNzM4LCJuYmYiOjE1MDE1OTAxMzgsImp0aSI6ImxFM1c1bzVIN1VSTzQ5U2EifQ.lsI26YI_528SKXyG9K189Nle-2HRLUpG2DOro6Aa6Oc"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "examples": {
                            "application/json": {
                                "error": {
                                    "status": 400,
                                    "title": "Bad Request",
                                    "detail": "Could not decode token: The token \"eyJ0eXAiOiJKV1QiLCJhsbGciOiJIUzI1NiJ9.eyJleHAiOjE1MTM2NzIwNDgsInN1YiI6NiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdC9hcGkvYXV0aGVudGljYXRlIiwiaWF0IjoxNTAxNTc2MDQ5LCJuYmYiOjE1MDE1NzYwNDksImp0aSI6IkZVcW42NXRlY0xDSm5kN1AifQ.K2wOp6Qk0URWaYV_FXxTqWQPx-zyyzglVryMpvLYZgU\" is an invalid JWS",
                                    "source": {
                                        "parameter": "http://localhost/api/authenticate/eyJ0eXAiOiJKV1QiLCJhsbGciOiJIUzI1NiJ9.eyJleHAiOjE1MTM2NzIwNDgsInN1YiI6NiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdC9hcGkvYXV0aGVudGljYXRlIiwiaWF0IjoxNTAxNTc2MDQ5LCJuYmYiOjE1MDE1NzYwNDksImp0aSI6IkZVcW42NXRlY0xDSm5kN1AifQ.K2wOp6Qk0URWaYV_FXxTqWQPx-zyyzglVryMpvLYZgU"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}